Make clang-tidy use mingw64
I give up on trying to make this platform-agnostic. There is seemingly
no way to just tell it to use gcc without having to specify an arch
or whether it's mingw. This fixes it for Windows (since Clang defaults
to MSVC), but it probably breaks it for Linux and others.

Also, I couldn't get --gcc-toolchain option to work through pre-commit,
so it's not necessarily going to use gcc 6.5.0. The local workaround
would be to ensure 6.5.0 is what's on discoverable on PATH.

I tried specifying a clang config file with --config as an extra arg,
but it doesn't respect that option through clang-tidy.