Store a single size field in the pcb_array struct
The original idea was to have count indicate the active processes and
capacity be the array size. The count could be used to validate process
selection. However, because inactive processes aren't necessarily at
the end of the array, this idea won't work. Therefore, only a single
field that stores the array size will suffice.