Exit if malloc fails to allocate process arrays
malloc could fail on any iteration and on any of the 3 fields. release()
doesn't know how up to which struct allocation succeeded and up to which
field of that struct it succeeded. It needs to know this because free()
is undefined for invalid pointers - and the rest of the pointers will
be garbage rather than NULL.

Exiting here avoids the default action in main() to run release() before
returning.